feat: support decimal TestCaseIterations and real-time adjusted hours UI updates
- Changed TestCaseIterations from int to decimal to support fractional
  iterations (e.g., 2.5) matching the Excel estimation spreadsheet
- Updated ProjectEntity model, MainViewModel calculation engine, and
  all test assertions for the decimal type change
- Tool now produces exact Excel match: Grand Total = 5262 with 2.5 iterations

UI Improvements:
- Changed all adjusted hours bindings from LostFocus to PropertyChanged
  for real-time derived task updates as user types
- Added AdjustedHoursConverter to handle empty/intermediate input states
  (backspace to empty → 0, partial input like '-' or '.' → 0)
- Derived Tasks panel now binds to effective totals (calculated + adjusted)
  instead of calculated-only values

Bug Fixes:
- Fixed SQL injection warning (EF1002) in DatabaseSeeder
- Fixed CS8618 nullable warning in EstimateDbContext
- Fixed static WeightedValues matrix corruption across test classes
- Fixed ProjectPersistenceTests double-component issue
- Fixed ConverterAdvancedTests SupportModules display name assertion

Test Coverage:
- Added ExcelVerificationTests (27 tests) validating all calculations
  against 'CO 23327 002 Final Estimate V1.0.xlsm'
- Added NegativeBoundaryTests (39 tests) for edge cases
- Added CoverageGapTests (83 tests) for remaining coverage gaps
- Added AdjustedHoursConverter tests (17 tests)
- Added per-type collaboration total tests (8 tests)
- Total: 776 tests passing, 0 warnings, 0 errors

Build:
- Suppressed NU1903 transitive vulnerability warning (SQLitePCLRaw)
- Disabled parallel test execution via xunit.runner.json

+++ InitialEstimatePOC/ViewModels/MainViewModel.cs
@@ -230,7 +230,7 @@ public partial class MainViewModel : ObservableObject
     private int _testCasesVeryComplex;
 
     [ObservableProperty]
-    private int _testCaseIterations = 1;
+    private decimal _testCaseIterations = 1m;
 
     // === Role Breakout ===
     [ObservableProperty]
@@ -504,7 +504,7 @@ public partial class MainViewModel : ObservableObject
     partial void OnTestCasesMediumChanged(int value) => Recalculate();
     partial void OnTestCasesComplexChanged(int value) => Recalculate();
     partial void OnTestCasesVeryComplexChanged(int value) => Recalculate();
-    partial void OnTestCaseIterationsChanged(int value) => Recalculate();
+    partial void OnTestCaseIterationsChanged(decimal value) => Recalculate();
     partial void OnTotalActualHoursChanged(decimal value) => Recalculate();
     partial void OnTimeForEstimatesChanged(decimal value) => Recalculate();
 
@@ -540,7 +540,7 @@ public partial class MainViewModel : ObservableObject
                                 + TestCasesComplex * r31Complex + TestCasesVeryComplex * r31VeryComplex;
             decimal defectHours = (TestCasesSimple * r32Simple + TestCasesMedium * r32Medium
                                 + TestCasesComplex * r32Complex + TestCasesVeryComplex * r32VeryComplex) * 0.1m;
-            SystemTestingHours = RoundUp((mainHours + defectHours) * Math.Max(1, TestCaseIterations));
+            SystemTestingHours = RoundUp((mainHours + defectHours) * Math.Max(1m, TestCaseIterations));
         }
         else
         {
